[feat] 타입 분리, 메시지 상수화 #173
