+++ src/constants/messages.ts
@@ -1,3 +1,11 @@
+export const ERROR_MESSAGES = {
+  INVALID_FILE_TYPE: '텍스트(.txt) 파일만 업로드할 수 있습니다.',
+  FILE_UPLOAD_FAILED: '파일 업로드에 실패했습니다. 다시 시도해주세요.',
+  EMPTY_FILE: '파일이 비어있습니다.',
+  NO_PROCESSABLE_TEXT: '처리할 수 있는 텍스트가 없습니다.',
+  UNKNOWN_ERROR: '알 수 없는 오류가 발생했습니다.',
+} as const;
+
 export const SIGNUP_VALIDATION = {
   REQUIRED: {
     EMAIL: '이메일을 입력해주세요.',
